Problem: tiles don't work well for listing issues
Firstly, there's a problem with fitting the title into a small
card, and how one would clip it.

Then, it still falls back to a list on mobiles / narrow browsers.

Solution: list issues in a list

Also, in order to avoid it being hard to scan, limit the entire
screen width to 1024px.